requirements.txt
confusable-homoglyphs==3.0.0
dj-database-url==0.5.0
Django==2.0.2
django-cors-headers==2.2.0
django-heroku==0.3.1
django-registration==2.4.1
django-webpack-loader==0.6.0
djangorestframework==3.7.3
djangorestframework-jwt==1.11.0
graphene==2.0.1
graphene-django==2.0.0
graphql-core==2.0
graphql-relay==0.4.5
gunicorn==19.7.1
iso8601==0.1.12
numpy==1.14.0
pandas==0.22.0
pbr==3.1.1
promise==2.1
psycopg2==2.7.4
PyJWT==1.6.0
python-dateutil==2.6.1
pytz==2017.3
Rx==1.6.1
singledispatch==3.4.0.3
six==1.11.0
stevedore==1.28.0
typing==3.6.4
virtualenv==15.1.0
virtualenv-clone==0.3.0
virtualenvwrapper==4.8.2
whitenoise==3.3.1
xlrd==1.1.0
